backend/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
certifi==2020.6.20
chardet==3.0.4
distlib==0.3.1
Django==3.1.5
django-cors-headers==3.7.0
djangorestframework==3.12.2
filelock==3.0.12
idna==2.10
pipenv==2020.6.2
pytz==2020.5
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
virtualenv==20.0.30
virtualenv-clone==0.5.4
wget==3.2
